📦 NEW: Transfer code from https://github.com/sguizard/nf-core-isoseq
The development started on my own repo. 
Since the code reach an operational state, I transfer it to the official Project Gene Switch repo.
The pipeline name will be renamed to stick with GS guidelines. This will be done at the update of nf-core templates to their last DSL2 version.